ActualStatus field replaces TranslationSucceeded ..,  TabItemConflicts.Fontweight used, Analyse conflicts without explicit priorities

+++ MainWindow.xaml.cs
@@ -1,10 +1,13 @@
 ﻿using Microsoft.Win32;
+
 using System;
 using System.Diagnostics;
 using System.IO;
+using System.Security.Policy;
 using System.Text;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Media;
 
 namespace grammlator {
    /// <summary>
@@ -14,10 +17,23 @@ namespace grammlator {
       public MainWindow()
       {
          InitializeComponent();
+
+         ActualStatus = new StatusStruct(StatusChanged); // Initial State: no flags are set
+         ActualStatus.SetAction(StatusChanged);
+
          Title = "grammlator - no file -";
-         SetInitiaStatus();
          FocusTextBox += HandleFocusTextBox;
          OnFocusTextBox(new FocusTextBoxEventArgs(SourceTextBox));
+
+         string SaveDirectory = SaveSourceOrResultDialog.InitialDirectory;
+         // do not offer the directory containing the assembly as default initial directory
+         if (SaveDirectory == GlobalVariables.AssemblyFullPath)
+            SaveSourceOrResultDialog.InitialDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+
+         string OpenSourceDirectory = OpenSourceFileDialog.InitialDirectory;
+         // do not offer the directory containing the assembly as default initial directory
+         if (OpenSourceDirectory == GlobalVariables.AssemblyFullPath)
+            OpenSourceFileDialog.InitialDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
       }
 
       const String fileFilter = "cs files (*.cs)|*.cs|All files (*.*)|*.*";
@@ -35,25 +51,40 @@ namespace grammlator {
          Filter = fileFilter
       };
 
-      
-      private bool SaveFile(string fileName, string title, string text)
+
+      private bool SaveFile(string lastFileName, string title, string text, bool allowSourceAsDestination)
       {
          Debug.Assert(!String.IsNullOrEmpty(title));
 
-         if (fileName != "")
+         if (lastFileName != "")
          {
-            SaveSourceOrResultDialog.InitialDirectory = Path.GetDirectoryName(fileName);
-            SaveSourceOrResultDialog.FileName = Path.GetFileName(fileName);
+            SaveSourceOrResultDialog.InitialDirectory = Path.GetDirectoryName(lastFileName);
+            SaveSourceOrResultDialog.FileName = Path.GetFileName(lastFileName);
          }
 
          SaveSourceOrResultDialog.Title = title;
          Boolean? r = SaveSourceOrResultDialog.ShowDialog();
          if (r != true)
             return false; // canceled by user
+
+         // Did the user select the source file?
+         bool SourceFileReplace = SourceFilename != "" && SourceFilename == SaveSourceOrResultDialog.FileName;
+
+         if (!allowSourceAsDestination && SourceFileReplace)
+         {
+            // ask user if source textbox contens and source file should be replaced by result
+            if (MessageBoxResult.Cancel == MessageBox.Show("Copy the result to the source text and save to source file?",
+               "Save result to source?", MessageBoxButton.OKCancel)
+               )
+               return false;
+            // Copy result to source to stay consistent
+            SourceTextBox.Text = ResultTextBox.Text;
+         }
+
+
          try
          {
             File.WriteAllText(SaveSourceOrResultDialog.FileName, text);
-            SetStatusHavingSourceFilename(SaveSourceOrResultDialog.FileName);
          }
          catch (Exception ex)
          {
@@ -61,61 +92,97 @@ namespace grammlator {
             return false;
          }
 
+         if (SourceFileReplace)
+         {
+            SourceFilename = SaveSourceOrResultDialog.FileName;
+            ActualStatus.SetFlags(StatusFlag.SourceHasFilename, true);
+            ActualStatus.SetFlags(StatusFlag.SourceTextChanged, false); // has just been saved
+         }
          return true;
       }
 
       private void MenuItemSaveSource_Click(Object sender, RoutedEventArgs e)
       {
-         SaveFile(SourceFilename, "Save source file", SourceTextBox.Text);
-         // SaveFile does: SetStatusHavingSourceFilename(SaveSourceOrResultDialog.FileName);
+         SaveSource();
+      }
+
+      private void MenuItemSaveResultToSource_Click(Object sender, RoutedEventArgs e)
+      {
+         if (!ActualStatus.TestFlags(StatusFlag.ResultAvailable))
+         {
+            // this should not occur because the "SaveResultToSource" should be disabled
+            MessageBox.Show("No result available");
+            return;
+         }
+
+         SourceTextBox.Text = ResultTextBox.Text;
+         SaveSource();
+      }
+
+      private void SaveSource()
+      {
+         if (!SaveFile(SourceFilename, "Save source file", SourceTextBox.Text, true))
+            return;
+
+         // Source has been saved (may be with a different filename)
+
+         // TODO mark sourcefile as unchanged (has been saved)
+
+         SourceFilename = SaveSourceOrResultDialog.FileName;
+         ActualStatus.SetFlags(StatusFlag.SourceHasFilename, true);
+         ActualStatus.SetFlags(StatusFlag.SourceTextChanged, false);
       }
 
       private void MenuItemSaveResult_Click(Object sender, RoutedEventArgs e)
       {
-         if (!TranslationSucceeded)
-            return; // Click should have been disabled, but be shure not to save incomplete result
+         if (!ActualStatus.TestFlags(StatusFlag.ResultAvailable))
+         {
+            // this should not occur because the "SaveResult" should be disabled
+            MessageBox.Show("No result available");
+            return;
+         };
 
+         // Preset ResultFilename with SourceFilename+"-generated"
          String ResultFilename = "";
-         if (!String.IsNullOrEmpty(SourceFilename))
+         if (SourceFilename != "")
             ResultFilename =
                 Path.GetDirectoryName(SourceFilename) + "\\" +
                 Path.GetFileNameWithoutExtension(SourceFilename) + "-generated" +
                 Path.GetExtension(SourceFilename);
 
-         SaveFile(ResultFilename, "Save result", ResultTextBox.Text); // TOCHECK: Has source been replaced?
+
+         if (!SaveFile(ResultFilename, "Save result", ResultTextBox.Text, false))
+            return;
+
+         // Result has been saved
 
       }
 
-      private void MenuItemSaveResultToSource_Click(Object sender, RoutedEventArgs e)
-      {
-         if (!TranslationSucceeded)
-            return; // Click should have been disabled, but be shure not to save incomplete result
-
-         SourceTextBox.Text = ResultTextBox.Text;
-         MenuItemSaveSource_Click(sender, e); // 
-      }
 
       private void MenuItemTranslateStandard_Click(Object _1, RoutedEventArgs _2)
          => Translate();
 
       private void MenuItemLoadSourceFile_Click(Object sender, RoutedEventArgs e)
       {
-         if (String.IsNullOrEmpty(SourceFilename))
-         {
-            OpenSourceFileDialog.FileName = "";
-         }
-         else
+         LoadSourcefile();
+      }
+
+      private Boolean LoadSourcefile()
+      {
+
+         OpenSourceFileDialog.FileName = "";
+
+         if (!String.IsNullOrEmpty(SourceFilename))
          {
             OpenSourceFileDialog.InitialDirectory = Path.GetDirectoryName(SourceFilename);
-            OpenSourceFileDialog.FileName = Path.GetFileName(SourceFilename);
          }
 
          OpenSourceFileDialog.Title = "Load source file";
          Boolean? result = OpenSourceFileDialog.ShowDialog();
          if (result != true)
-            return; // canceled by user
+            return false; // canceled by user
 
-         ClearResultsAndReadFileToSourceTextbox();
+         return ClearResultsAndReadFileToSourceTextbox(); // and set ActualStatus
       }
 
       private void MenuItemReloadAndTranslate_Click(Object sender, RoutedEventArgs e)
@@ -123,11 +190,11 @@ namespace grammlator {
          if (String.IsNullOrEmpty(SourceFilename))
          {
             MessageBox.Show("unknown filename");
+            return;
          }
-         else if (ClearResultsAndReadFileToSourceTextbox())
-         {
+
+         if (LoadSourcefile())
             Translate();
-         }
       }
 
       private void CompareIgnoringSeparators_Click(Object _1, RoutedEventArgs _2)
@@ -257,13 +324,8 @@ namespace grammlator {
 
       private void SourceTextBox_TextChanged(object sender, TextChangedEventArgs args)
       {
-         if (MenuItemTranslateStandard == null)
-            return;
-
-         if ((sender as TextBox)?.LineCount > 10)
-            MenuItemTranslateStandard.IsEnabled = true;
-         else
-            MenuItemTranslateStandard.IsEnabled = false;
+         ActualStatus.SetFlags(StatusFlag.SourceTextChanged, true);
+         ActualStatus.SetFlags(StatusFlag.SourceNotEmpty, (sender as TextBox)?.LineCount > 10);
       }
    }
 }
